[feat] Quiz card & seperated atoms (ONE-87)
