+++ src/types.ts
@@ -1,0 +1,4 @@
+export type Answer = {
+    question: number;
+    answer: number;
+}
